Project_7 Phone and Table Mockups
The small images/headlines above the hero image are links to other
(possible audio) stories. I’m not sure how to deal with them. On the
phone, it probably would be best just to scroll down. On the tablet,
there’s room for them to the right of the hero image—I don’t know how
best to use that real estate. I don’t have a tablet but my guess is
that more actual reading can happen there. I put the links to
blog/books/teaching/contact into the hamburger, and don’t know if I
will be up to the coding challenge there.